+++ .dockerignore
@@ -1,0 +1,7 @@
+__pycache__/
+.git/
+.env
+*.pyc
+*.pyo
+*.pyd
+*.zip
