+++ docker-compose.yml
@@ -1,0 +1,33 @@
+version: "3.8"
+
+services:
+  fastapi:
+    build: .
+    container_name: fastapi-container
+    ports:
+      - "8000:8000"
+    # volumes:
+    #   - ./app:/app
+    env_file:
+      - .env # 다른 경로의 .env 파일 지정
+    environment:
+      - PERPLEXITY_API_KEY=${PERPLEXITY_API_KEY}
+      - OPENAI_API_KEY=${OPENAI_API_KEY}
+      - ANTHROPIC_API_KEY=${ANTHROPIC_API_KEY}
+
+      # - ENV=production
+  #     - AWS_ACCESS_KEY_ID=test
+  #     - AWS_SECRET_ACCESS_KEY=test
+  #     - DYNAMODB_ENDPOINT=http://dynamodb:8000  # DynamoDB 컨테이너를 참조
+
+  # dynamodb:
+  #   image: amazon/dynamodb-local
+  #   container_name: dynamodb
+  #   ports:
+  #     - "8000:8000"
+  #   environment:
+  #     - AWS_ACCESS_KEY_ID=test
+  #     - AWS_SECRET_ACCESS_KEY=test
+  #   command: "-jar DynamoDBLocal.jar -sharedDb"
+  #   volumes:
+  #     - ./dynamodb:/data  # 데이터 지속성을 위한 볼륨 마운트
