+++ requirements.txt
@@ -1,0 +1,39 @@
+annotated-types==0.7.0
+anthropic==0.42.0
+anyio==4.8.0
+certifi==2024.12.14
+charset-normalizer==3.4.1
+defusedxml==0.7.1
+distro==1.9.0
+fastapi==0.115.6
+git-filter-repo==2.47.0
+h11==0.14.0
+httpcore==1.0.7
+httpx==0.28.1
+idna==3.10
+jiter==0.8.2
+jsonpatch==1.33
+jsonpointer==3.0.0
+langchain-anthropic==0.3.1
+langchain-core==0.3.29
+langchain-openai==0.3.0
+langsmith==0.2.10
+mangum==0.19.0
+openai==1.59.6
+orjson==3.10.14
+packaging==24.2
+pydantic==2.10.5
+pydantic_core==2.27.2
+python-dotenv==1.0.1
+PyYAML==6.0.2
+regex==2024.11.6
+requests==2.32.3
+requests-toolbelt==1.0.0
+sniffio==1.3.1
+starlette==0.41.3
+tenacity==9.0.0
+tiktoken==0.8.0
+tqdm==4.67.1
+typing_extensions==4.12.2
+urllib3==2.3.0
+uvicorn
