--- seteuk_basic_app/.github/workflows/docker-build.yml
@@ -1,38 +1,0 @@
-name: Build and Push Docker Image
-
-on:
-  push:
-    branches:
-      - dev
-      - main
-
-jobs:
-  build-and-deploy:
-    runs-on: ubuntu-latest
-
-    steps:
-      # 1. GitHub 저장소 클론
-      - name: Checkout code
-        uses: actions/checkout@v3
-
-      # 2. AWS 로그인
-      - name: Log in to AWS ECR
-        id: login-ecr
-        uses: aws-actions/amazon-ecr-login@v1
-
-      # 3. Docker 이미지 빌드
-      - name: Build Docker image
-        run: |
-          docker build -t ${{ secrets.AWS_ACCOUNT_ID }}.dkr.ecr.${{ secrets.AWS_REGION }}.amazonaws.com/lambda-fastapi:${{ github.sha }} .
-
-      # 4. Docker 이미지 푸시
-      - name: Push Docker image to ECR
-        run: |
-          docker push ${{ secrets.AWS_ACCOUNT_ID }}.dkr.ecr.${{ secrets.AWS_REGION }}.amazonaws.com/lambda-fastapi:${{ github.sha }}
-
-      # 5. Lambda 함수 업데이트
-      - name: Deploy Lambda function
-        run: |
-          aws lambda update-function-code \
-            --function-name lambdaProto \
-            --image-uri ${{ secrets.AWS_ACCOUNT_ID }}.dkr.ecr.${{ secrets.AWS_REGION }}.amazonaws.com/lambda-fastapi:${{ github.sha }}
